test-requirements.txt
pytest>=3.0
requests
PyYaml